Changement total des retours vers Response Entity dans le controller Utilisateur avec le throw des exceptions dans le service

+++ src/main/java/com/youcode/rentalhive/service/utilisateur/implemetation/UtilisateurServiceImpl.java
@@ -1,5 +1,7 @@
     package com.youcode.rentalhive.service.utilisateur.implemetation;
 
+    import com.youcode.rentalhive.handler.exception.OperationException;
+    import com.youcode.rentalhive.handler.exception.ResourceNotFoundException;
     import com.youcode.rentalhive.model.entity.Utilisateur;
     import com.youcode.rentalhive.repository.UtilisateurRepository;
     import com.youcode.rentalhive.service.utilisateur.UtilisateurService;
@@ -18,12 +20,16 @@
 
         @Override
         public Utilisateur createUtilisateur(Utilisateur utilisateur) {
+            if(utilisateurRepository.existsByEmail(utilisateur.getEmail())){
+                throw new OperationException("L'email fourni est déjà utilisé");
+            }
             return utilisateurRepository.save(utilisateur);
         }
 
         @Override
-        public Optional<Utilisateur> getUtilisateurById(Long id) {
-            return utilisateurRepository.findById(id);
+        public Utilisateur getUtilisateurById(Long id) {
+            return utilisateurRepository.findById(id)
+                    .orElseThrow(() -> new ResourceNotFoundException("L'utilisateur avec l'id : " + id + " n'existe pas"));
         }
 
         @Override
@@ -39,17 +45,31 @@
         @Override
         public Utilisateur updateUtilisateur(Utilisateur utilisateur, Long id) {
             Optional<Utilisateur> existingUtilisateurOptional = utilisateurRepository.findById(id);
+
             if (existingUtilisateurOptional.isPresent()) {
                 Utilisateur existingUtilisateur = existingUtilisateurOptional.get();
+                if (!existingUtilisateur.getEmail().equals(utilisateur.getEmail()) &&
+                        utilisateurRepository.existsByEmail(utilisateur.getEmail())) {
+                    throw new OperationException("L'email fourni est déjà utilisé par un autre utilisateur.");
+                }
                 existingUtilisateur.setNomUtilisateur(utilisateur.getNomUtilisateur());
                 existingUtilisateur.setEmail(utilisateur.getEmail());
+
                 return utilisateurRepository.save(existingUtilisateur);
+            } else {
+                throw new ResourceNotFoundException("L'utilisateur avec l'ID : " + id + " n'existe pas.");
             }
-            return null;
         }
 
         @Override
         public void deleteUtilisateur(Long id) {
-            utilisateurRepository.deleteById(id);
+            Optional<Utilisateur> existingUtilisateurOptional = utilisateurRepository.findById(id);
+
+            if (existingUtilisateurOptional.isPresent()) {
+                utilisateurRepository.deleteById(id);
+            } else {
+                throw new ResourceNotFoundException("L'utilisateur avec l'ID : " + id + " n'existe pas.");
+            }
         }
+
     }
